Refactor IRacingTelemetryService into partial classes
